requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.3.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
outcome==1.3.0.post0
packaging==24.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
regex==2023.12.25
requests==2.31.0
scrapetube==2.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
